setup NoteScreen

+++ app/src/main/java/com/fyp/inotes/feature_note/domain/model/Note.kt
@@ -12,7 +12,7 @@ import com.fyp.inotes.ui.theme.Violet
 data class Note(
     val title: String,
     val content: String,
-    val color: Int,
+    val color: Long,
     val timeStamp: Long,
     @PrimaryKey val id: Int? = null
 ) {
